+++ users/src/main/java/ru/yandex/practicum/filmorate/films/domain/port/UpdateUserCommand.java
@@ -2,7 +2,7 @@ package ru.yandex.practicum.filmorate.films.domain.port;
 
 import java.time.LocalDate;
 
-public record UpdateUserCommand(Integer id,
+public record UpdateUserCommand(Long id,
                                 String email,
                                 String login,
                                 String name,
