+++ users/src/main/java/ru/yandex/practicum/filmorate/films/web/dto/UserResponse.java
@@ -2,7 +2,7 @@ package ru.yandex.practicum.filmorate.films.web.dto;
 
 import java.time.LocalDate;
 
-public record UserResponse(Integer id,
+public record UserResponse(Long id,
                            String email,
                            String login,
                            String name,
